rename account_dialog module to account_activity
